Reorganized constraints files.
